Saturated PSF alignment script
Code to align two images in x, and y for saturated PSF.

It uses the aery rings to find the best matching x,y values and shift
the images. To do this, the code creates an annular ring (size can be
adjusted appropriately), and an initial guess of the stellar center
(within +- 5px).

To run use the following command:
python pyAlign.py (or ./pyAlign.py if made executable)